Add files via upload
Polymerase prototype 1.2: The movement of the cutout for the insertion of the plus plus piece works well, but somehow in the process the groove/notch that allows the polymerase model to fit over the PP DNA strand was shortened by 1 or 2 mm. It needs to be 20 mm, but it is 19 mm. This was adjusted for prototype 1.3.